# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationDetail": {
        "additionalProperties": false,
        "properties": {
          "allowedActions": {
            "items": {
              "enum": [
                "cancel",
                "retry"
              ],
              "type": "string"
            },
            "title": "Allowedactions",
            "type": "array"
          },
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormessage"
          },
          "finishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "heartbeatAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeatat"
          },
          "maxAttempts": {
            "title": "Maxattempts",
            "type": "integer"
          },
          "operationId": {
            "format": "uuid",
            "title": "Operationid",
            "type": "string"
          },
          "progressCurrent": {
            "title": "Progresscurrent",
            "type": "integer"
          },
          "progressTotal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progresstotal"
          },
          "progressUnit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progressunit"
          },
          "queuedAt": {
            "format": "date-time",
            "title": "Queuedat",
            "type": "string"
          },
          "resultSummary": {
            "additionalProperties": true,
            "title": "Resultsummary",
            "type": "object"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "stageCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stagecode"
          },
          "stageLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stagelabel"
          },
          "startedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "status": {
            "$ref": "#/components/schemas/OperationStatus"
          },
          "statusUrl": {
            "title": "Statusurl",
            "type": "string"
          },
          "targetId": {
            "format": "uuid",
            "title": "Targetid",
            "type": "string"
          },
          "targetType": {
            "title": "Targettype",
            "type": "string"
          },
          "taskType": {
            "title": "Tasktype",
            "type": "string"
          },
          "warningCount": {
            "title": "Warningcount",
            "type": "integer"
          }
        },
        "required": [
          "operationId",
          "status",
          "statusUrl",
          "targetType",
          "targetId",
          "taskType",
          "progressCurrent",
          "attempt",
          "maxAttempts",
          "warningCount",
          "retryable",
          "resultSummary",
          "queuedAt",
          "allowedActions"
        ],
        "title": "OperationDetail",
        "type": "object"
      },
      "OperationPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OperationDetail"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "title": "OperationPage",
        "type": "object"
      },
      "OperationStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "partial_succeeded",
          "failed",
          "cancelled"
        ],
        "title": "OperationStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "企业 RAG 知识库 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/operations": {
      "get": {
        "operationId": "operationsList",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "taskType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tasktype"
            }
          },
          {
            "in": "query",
            "name": "targetType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Targettype"
            }
          },
          {
            "in": "query",
            "name": "targetId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Targetid"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "enum": [
                20,
                50,
                100
              ],
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "-queued_at",
              "enum": [
                "queued_at",
                "-queued_at",
                "created_at",
                "-created_at"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "rag_admin_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rag Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Operations",
        "tags": [
          "异步任务"
        ]
      }
    },
    "/api/v1/operations/{operationId}": {
      "get": {
        "operationId": "operationsGet",
        "parameters": [
          {
            "in": "path",
            "name": "operationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operationid",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "rag_admin_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rag Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Operation",
        "tags": [
          "异步任务"
        ]
      }
    },
    "/api/v1/operations/{operationId}:cancel": {
      "post": {
        "operationId": "operationsCancel",
        "parameters": [
          {
            "in": "path",
            "name": "operationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operationid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "rag_admin_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rag Admin Access"
            }
          },
          {
            "in": "cookie",
            "name": "rag_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rag Csrf"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Operation",
        "tags": [
          "异步任务"
        ]
      }
    },
    "/api/v1/operations/{operationId}:retry": {
      "post": {
        "operationId": "operationsRetry",
        "parameters": [
          {
            "in": "path",
            "name": "operationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operationid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "rag_admin_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rag Admin Access"
            }
          },
          {
            "in": "cookie",
            "name": "rag_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rag Csrf"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Operation",
        "tags": [
          "异步任务"
        ]
      }
    }
  }
}